refactor(service-client): Make service client easy to mock
