+++ iothub/service/src/ServiceClient.cs
@@ -2,8 +2,16 @@
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
 using System;
+using System.Collections.Generic;
+using System.Net;
+using System.Net.Http;
 using System.Threading;
 using System.Threading.Tasks;
+using Microsoft.Azure.Amqp;
+using Microsoft.Azure.Amqp.Framing;
+using Microsoft.Azure.Devices.Common;
+using Microsoft.Azure.Devices.Common.Data;
+using Microsoft.Azure.Devices.Common.Exceptions;
 using Microsoft.Azure.Devices.Shared;
 
 namespace Microsoft.Azure.Devices
@@ -32,14 +40,51 @@ namespace Microsoft.Azure.Devices
     /// Contains methods that services can use to send messages to devices/modules,
     /// invoke a direct method on a device/module and deliver notifications for file upload and cloud-to-device operations.
     /// </summary>
-    public abstract class ServiceClient : IDisposable
+    public class ServiceClient : IDisposable
     {
+        private const string _statisticsUriFormat = "/statistics/service?" + ClientApiVersionHelper.ApiVersionQueryString;
+        private const string _purgeMessageQueueFormat = "/devices/{0}/commands?" + ClientApiVersionHelper.ApiVersionQueryString;
+        private const string _deviceMethodUriFormat = "/twins/{0}/methods?" + ClientApiVersionHelper.ApiVersionQueryString;
+        private const string _moduleMethodUriFormat = "/twins/{0}/modules/{1}/methods?" + ClientApiVersionHelper.ApiVersionQueryString;
+
+        private static readonly TimeSpan s_defaultOperationTimeout = TimeSpan.FromSeconds(100);
+
+        private readonly FaultTolerantAmqpObject<SendingAmqpLink> _faultTolerantSendingLink;
+        private readonly string _sendingPath;
+        private readonly AmqpFeedbackReceiver _feedbackReceiver;
+        private readonly AmqpFileNotificationReceiver _fileNotificationReceiver;
+        private readonly IHttpClientHelper _httpClientHelper;
+        private readonly string _iotHubName;
+        private readonly ServiceClientOptions _clientOptions;
+        private readonly TimeSpan _openTimeout;
+        private readonly TimeSpan _operationTimeout;
+
+        private int _sendingDeliveryTag;
+
+        internal readonly IotHubConnection Connection;
+
         /// <summary>
-        /// Make this constructor internal so that only this library may implement this abstract class.
+        /// Creates an instance of <see cref="ServiceClient"/>.
         /// </summary>
-        internal ServiceClient()
+        public ServiceClient()
         {
-            TlsVersions.Instance.SetLegacyAcceptableVersions();
+        }
+
+        internal ServiceClient(IotHubConnection connection, string iotHubName, IHttpClientHelper httpClientHelper, ServiceClientOptions options)
+        {
+            Connection = connection;
+            _iotHubName = iotHubName;
+            _clientOptions = options;
+            _httpClientHelper = httpClientHelper;
+            _openTimeout = IotHubConnection.DefaultOpenTimeout;
+            _operationTimeout = IotHubConnection.DefaultOperationTimeout;
+            _sendingPath = "/messages/deviceBound";
+            _faultTolerantSendingLink = new FaultTolerantAmqpObject<SendingAmqpLink>(CreateSendingLinkAsync, Connection.CloseLink);
+            _feedbackReceiver = new AmqpFeedbackReceiver(Connection);
+            _fileNotificationReceiver = new AmqpFileNotificationReceiver(Connection);
+
+            // Set the trace provider for the AMQP library.
+            AmqpTrace.Provider = new AmqpTransportLog();
         }
 
         /// <summary>
@@ -64,7 +109,17 @@ namespace Microsoft.Azure.Devices
         /// Releases unmanaged and - optionally - managed resources.
         /// </summary>
         /// <param name="disposing"><c>true</c> to release both managed and unmanaged resources; <c>false</c> to release only unmanaged resources.</param>
-        protected virtual void Dispose(bool disposing) { }
+        protected virtual void Dispose(bool disposing)
+        {
+            if (disposing)
+            {
+                _faultTolerantSendingLink.Dispose();
+                _fileNotificationReceiver.Dispose();
+                _feedbackReceiver.Dispose();
+                Connection.Dispose();
+                _httpClientHelper.Dispose();
+            }
+        }
 
         /// <summary>
         /// Create an instance of ServiceClient from the specified IoT Hub connection string using specified Transport Type.
@@ -95,55 +150,184 @@ namespace Microsoft.Azure.Devices
 
             var iotHubConnectionString = IotHubConnectionString.Parse(connectionString);
             bool useWebSocketOnly = transportType == TransportType.Amqp_WebSocket_Only;
-            var serviceClient = new AmqpServiceClient(iotHubConnectionString, useWebSocketOnly, transportSettings, options);
+            var iotHubConnection = new IotHubConnection(iotHubConnectionString, AccessRights.ServiceConnect, useWebSocketOnly, transportSettings);
+            var httpClientHelper = new HttpClientHelper(
+               iotHubConnectionString.HttpsEndpoint,
+               iotHubConnectionString,
+               ExceptionHandlingHelper.GetDefaultErrorMapping(),
+               s_defaultOperationTimeout,
+               transportSettings.HttpProxy);
+            var serviceClient = new ServiceClient(iotHubConnection, iotHubConnectionString.IotHubName, httpClientHelper, options);
             return serviceClient;
         }
 
         /// <summary>
-        /// Open the ServiceClient instance.
+        /// Open the ServiceClient instance. This call is made over AMQP.
         /// </summary>
-        public abstract Task OpenAsync();
+        public virtual async Task OpenAsync()
+        {
+            Logging.Enter(this, $"Opening AmqpServiceClient", nameof(OpenAsync));
+
+            await _faultTolerantSendingLink.OpenAsync(_openTimeout).ConfigureAwait(false);
+            await _feedbackReceiver.OpenAsync().ConfigureAwait(false);
+
+            Logging.Exit(this, $"Opening AmqpServiceClient", nameof(OpenAsync));
+        }
 
         /// <summary>
-        /// Close the ServiceClient instance.
+        /// Close the ServiceClient instance. This call is made over AMQP.
         /// </summary>
-        public abstract Task CloseAsync();
+        public virtual async Task CloseAsync()
+        {
+            Logging.Enter(this, $"Closing AmqpServiceClient", nameof(CloseAsync));
+
+            await _faultTolerantSendingLink.CloseAsync().ConfigureAwait(false);
+            await _feedbackReceiver.CloseAsync().ConfigureAwait(false);
+            await _fileNotificationReceiver.CloseAsync().ConfigureAwait(false);
+            await Connection.CloseAsync().ConfigureAwait(false);
+
+            Logging.Exit(this, $"Closing AmqpServiceClient", nameof(CloseAsync));
+        }
 
         /// <summary>
-        /// Send a cloud-to-device message to the specified device.
+        /// Send a cloud-to-device message to the specified device. This call is made over AMQP.
         /// </summary>
         /// <param name="deviceId">The device identifier for the target device.</param>
         /// <param name="message">The cloud-to-device message.</param>
         /// <param name="timeout">The operation timeout, which defaults to 1 minute if unspecified.</param>
-        public abstract Task SendAsync(string deviceId, Message message, TimeSpan? timeout = null);
+        public virtual async Task SendAsync(string deviceId, Message message, TimeSpan? timeout = null)
+        {
+            Logging.Enter(this, $"Sending message with Id [{message?.MessageId}] for device {deviceId}", nameof(SendAsync));
+
+            if (string.IsNullOrWhiteSpace(deviceId))
+            {
+                throw new ArgumentNullException(nameof(deviceId));
+            }
+
+            if (message == null)
+            {
+                throw new ArgumentNullException(nameof(message));
+            }
+
+            if (_clientOptions?.SdkAssignsMessageId == SdkAssignsMessageId.WhenUnset && message.MessageId == null)
+            {
+                message.MessageId = Guid.NewGuid().ToString();
+            }
+
+            if (message.IsBodyCalled)
+            {
+                message.ResetBody();
+            }
+
+            timeout ??= _operationTimeout;
+
+            using AmqpMessage amqpMessage = MessageConverter.MessageToAmqpMessage(message);
+            amqpMessage.Properties.To = "/devices/" + WebUtility.UrlEncode(deviceId) + "/messages/deviceBound";
+
+            try
+            {
+                SendingAmqpLink sendingLink = await GetSendingLinkAsync().ConfigureAwait(false);
+                Outcome outcome = await sendingLink
+                    .SendMessageAsync(amqpMessage, IotHubConnection.GetNextDeliveryTag(ref _sendingDeliveryTag), AmqpConstants.NullBinary, timeout.Value)
+                    .ConfigureAwait(false);
+
+                Logging.Info(this, $"Outcome was: {outcome?.DescriptorName}", nameof(SendAsync));
+
+                if (outcome.DescriptorCode != Accepted.Code)
+                {
+                    throw AmqpErrorMapper.GetExceptionFromOutcome(outcome);
+                }
+            }
+            catch (Exception ex) when (!(ex is TimeoutException) && !ex.IsFatal())
+            {
+                Logging.Error(this, $"{nameof(SendAsync)} threw an exception: {ex}", nameof(SendAsync));
+                throw AmqpClientHelper.ToIotHubClientContract(ex);
+            }
+            finally
+            {
+                Logging.Exit(this, $"Sending message [{message?.MessageId}] for device {deviceId}", nameof(SendAsync));
+            }
+        }
 
         /// <summary>
-        /// Removes all cloud-to-device messages from a device's queue.
+        /// Removes all cloud-to-device messages from a device's queue. This call is made over HTTP.
         /// </summary>
         /// <param name="deviceId">The device identifier for the target device.</param>
         /// <param name="cancellationToken">A cancellation token to cancel the operation.</param>
-        public abstract Task<PurgeMessageQueueResult> PurgeMessageQueueAsync(string deviceId, CancellationToken cancellationToken = default);
+        public virtual Task<PurgeMessageQueueResult> PurgeMessageQueueAsync(string deviceId, CancellationToken cancellationToken = default)
+        {
+            Logging.Enter(this, $"Purging message queue for device: {deviceId}", nameof(PurgeMessageQueueAsync));
+
+            try
+            {
+                var errorMappingOverrides = new Dictionary<HttpStatusCode, Func<HttpResponseMessage, Task<Exception>>>
+                {
+                    { HttpStatusCode.NotFound, responseMessage => Task.FromResult((Exception)new DeviceNotFoundException(deviceId)) }
+                };
+
+                return _httpClientHelper.DeleteAsync<PurgeMessageQueueResult>(GetPurgeMessageQueueAsyncUri(deviceId), errorMappingOverrides, null, cancellationToken);
+            }
+            catch (Exception ex)
+            {
+                Logging.Error(this, $"{nameof(PurgeMessageQueueAsync)} threw an exception: {ex}", nameof(PurgeMessageQueueAsync));
+                throw;
+            }
+            finally
+            {
+                Logging.Exit(this, $"Purging message queue for device: {deviceId}", nameof(PurgeMessageQueueAsync));
+            }
+        }
 
         /// <summary>
         /// Get the <see cref="FeedbackReceiver{FeedbackBatch}"/> which can deliver acknowledgments for messages sent to a device/module from IoT Hub.
+        /// This call is made over AMQP.
         /// For more information see <see href="https://docs.microsoft.com/en-us/azure/iot-hub/iot-hub-devguide-messages-c2d#message-feedback"/>.
         /// </summary>
         /// <returns>An instance of <see cref="FeedbackReceiver{FeedbackBatch}"/>.</returns>
-        public abstract FeedbackReceiver<FeedbackBatch> GetFeedbackReceiver();
+        public virtual FeedbackReceiver<FeedbackBatch> GetFeedbackReceiver()
+        {
+            return _feedbackReceiver;
+        }
 
         /// <summary>
         /// Get the <see cref="FileNotificationReceiver{FileNotification}"/> which can deliver notifications for file upload operations.
+        /// This call is made over AMQP.
         /// For more information see <see href = "https://docs.microsoft.com/en-us/azure/iot-hub/iot-hub-devguide-file-upload#file-upload-notifications"/>.
         /// </summary>
         /// <returns>An instance of <see cref="FileNotificationReceiver{FileNotification}"/>.</returns>
-        public abstract FileNotificationReceiver<FileNotification> GetFileNotificationReceiver();
+        public virtual FileNotificationReceiver<FileNotification> GetFileNotificationReceiver()
+        {
+            return _fileNotificationReceiver;
+        }
 
         /// <summary>
-        /// Gets service statistics for the IoT Hub.
+        /// Gets service statistics for the IoT Hub. This call is made over HTTP.
         /// </summary>
         /// <param name="cancellationToken">A cancellation token to cancel the operation.</param>
         /// <returns>The service statistics that can be retrieved from IoT Hub, eg. the number of devices connected to the hub.</returns>
-        public abstract Task<ServiceStatistics> GetServiceStatisticsAsync(CancellationToken cancellationToken = default);
+        public virtual Task<ServiceStatistics> GetServiceStatisticsAsync(CancellationToken cancellationToken = default)
+        {
+            Logging.Enter(this, $"Getting service statistics", nameof(GetServiceStatisticsAsync));
+
+            try
+            {
+                var errorMappingOverrides = new Dictionary<HttpStatusCode, Func<HttpResponseMessage, Task<Exception>>>
+                {
+                    { HttpStatusCode.NotFound, responseMessage => Task.FromResult((Exception)new IotHubNotFoundException(_iotHubName)) }
+                };
+
+                return _httpClientHelper.GetAsync<ServiceStatistics>(GetStatisticsUri(), errorMappingOverrides, null, cancellationToken);
+            }
+            catch (Exception ex)
+            {
+                Logging.Error(this, $"{nameof(GetServiceStatisticsAsync)} threw an exception: {ex}", nameof(GetServiceStatisticsAsync));
+                throw;
+            }
+            finally
+            {
+                Logging.Exit(this, $"Getting service statistics", nameof(GetServiceStatisticsAsync));
+            }
+        }
 
         /// <summary>
         /// Interactively invokes a method on a device.
@@ -152,7 +336,10 @@ namespace Microsoft.Azure.Devices
         /// <param name="cloudToDeviceMethod">Parameters to execute a direct method on the device.</param>
         /// <param name="cancellationToken">A cancellation token to cancel the operation.</param>
         /// <returns>The <see cref="CloudToDeviceMethodResult"/>.</returns>
-        public abstract Task<CloudToDeviceMethodResult> InvokeDeviceMethodAsync(string deviceId, CloudToDeviceMethod cloudToDeviceMethod, CancellationToken cancellationToken = default);
+        public virtual Task<CloudToDeviceMethodResult> InvokeDeviceMethodAsync(string deviceId, CloudToDeviceMethod cloudToDeviceMethod, CancellationToken cancellationToken = default)
+        {
+            return InvokeDeviceMethodAsync(GetDeviceMethodUri(deviceId), cloudToDeviceMethod, cancellationToken);
+        }
 
         /// <summary>
         /// Interactively invokes a method on a module.
@@ -162,7 +349,20 @@ namespace Microsoft.Azure.Devices
         /// <param name="cloudToDeviceMethod">Parameters to execute a direct method on the module.</param>
         /// <param name="cancellationToken">A cancellation token to cancel the operation.</param>
         /// <returns>The <see cref="CloudToDeviceMethodResult"/>.</returns>
-        public abstract Task<CloudToDeviceMethodResult> InvokeDeviceMethodAsync(string deviceId, string moduleId, CloudToDeviceMethod cloudToDeviceMethod, CancellationToken cancellationToken = default);
+        public virtual Task<CloudToDeviceMethodResult> InvokeDeviceMethodAsync(string deviceId, string moduleId, CloudToDeviceMethod cloudToDeviceMethod, CancellationToken cancellationToken = default)
+        {
+            if (string.IsNullOrWhiteSpace(deviceId))
+            {
+                throw new ArgumentNullException(nameof(deviceId));
+            }
+
+            if (string.IsNullOrWhiteSpace(moduleId))
+            {
+                throw new ArgumentNullException(nameof(moduleId));
+            }
+
+            return InvokeDeviceMethodAsync(GetModuleMethodUri(deviceId, moduleId), cloudToDeviceMethod, cancellationToken);
+        }
 
         /// <summary>
         /// Send a cloud-to-device message to the specified module.
@@ -171,6 +371,155 @@ namespace Microsoft.Azure.Devices
         /// <param name="moduleId">The module identifier for the target module.</param>
         /// <param name="message">The cloud-to-module message.</param>
         ///  <param name="timeout">The operation timeout, which defaults to 1 minute if unspecified.</param>
-        public abstract Task SendAsync(string deviceId, string moduleId, Message message, TimeSpan? timeout = null);
+        public virtual async Task SendAsync(string deviceId, string moduleId, Message message, TimeSpan? timeout = null)
+        {
+            Logging.Enter(this, $"Sending message with Id [{message?.MessageId}] for device {deviceId}, module {moduleId}", nameof(SendAsync));
+
+            if (string.IsNullOrWhiteSpace(deviceId))
+            {
+                throw new ArgumentNullException(nameof(deviceId));
+            }
+
+            if (string.IsNullOrWhiteSpace(moduleId))
+            {
+                throw new ArgumentNullException(nameof(moduleId));
+            }
+
+            if (message == null)
+            {
+                throw new ArgumentNullException(nameof(message));
+            }
+
+            if (_clientOptions?.SdkAssignsMessageId == SdkAssignsMessageId.WhenUnset && message.MessageId == null)
+            {
+                message.MessageId = Guid.NewGuid().ToString();
+            }
+
+            if (message.IsBodyCalled)
+            {
+                message.ResetBody();
+            }
+
+            timeout ??= _operationTimeout;
+
+            using AmqpMessage amqpMessage = MessageConverter.MessageToAmqpMessage(message);
+            amqpMessage.Properties.To = "/devices/" + WebUtility.UrlEncode(deviceId) + "/modules/" + WebUtility.UrlEncode(moduleId) + "/messages/deviceBound";
+            try
+            {
+                SendingAmqpLink sendingLink = await GetSendingLinkAsync().ConfigureAwait(false);
+                Outcome outcome = await sendingLink
+                    .SendMessageAsync(
+                        amqpMessage,
+                        IotHubConnection.GetNextDeliveryTag(ref _sendingDeliveryTag),
+                        AmqpConstants.NullBinary,
+                        timeout.Value)
+                    .ConfigureAwait(false);
+
+                Logging.Info(this, $"Outcome was: {outcome?.DescriptorName}", nameof(SendAsync));
+
+                if (outcome.DescriptorCode != Accepted.Code)
+                {
+                    throw AmqpErrorMapper.GetExceptionFromOutcome(outcome);
+                }
+            }
+            catch (Exception ex) when (!ex.IsFatal())
+            {
+                Logging.Error(this, $"{nameof(SendAsync)} threw an exception: {ex}", nameof(SendAsync));
+                throw AmqpClientHelper.ToIotHubClientContract(ex);
+            }
+            finally
+            {
+                Logging.Exit(this, $"Sending message with Id [{message?.MessageId}] for device {deviceId}, module {moduleId}", nameof(SendAsync));
+            }
+        }
+
+        private Task<SendingAmqpLink> CreateSendingLinkAsync(TimeSpan timeout)
+        {
+            return Connection.CreateSendingLinkAsync(_sendingPath, timeout);
+        }
+
+        private async Task<SendingAmqpLink> GetSendingLinkAsync()
+        {
+            Logging.Enter(this, $"_faultTolerantSendingLink = {_faultTolerantSendingLink?.GetHashCode()}", nameof(GetSendingLinkAsync));
+
+            try
+            {
+                if (!_faultTolerantSendingLink.TryGetOpenedObject(out SendingAmqpLink sendingLink))
+                {
+                    sendingLink = await _faultTolerantSendingLink.GetOrCreateAsync(_openTimeout).ConfigureAwait(false);
+                }
+
+                Logging.Info(this, $"Retrieved SendingAmqpLink [{sendingLink?.Name}]", nameof(GetSendingLinkAsync));
+
+                return sendingLink;
+            }
+            finally
+            {
+                Logging.Exit(this, $"_faultTolerantSendingLink = {_faultTolerantSendingLink?.GetHashCode()}", nameof(GetSendingLinkAsync));
+            }
+        }
+
+        private Task<CloudToDeviceMethodResult> InvokeDeviceMethodAsync(Uri uri,
+            CloudToDeviceMethod cloudToDeviceMethod,
+            CancellationToken cancellationToken)
+        {
+            Logging.Enter(this, $"Invoking device method for: {uri}", nameof(InvokeDeviceMethodAsync));
+
+            try
+            {
+                TimeSpan timeout = GetInvokeDeviceMethodOperationTimeout(cloudToDeviceMethod);
+
+                return _httpClientHelper.PostAsync<CloudToDeviceMethod, CloudToDeviceMethodResult>(
+                    uri,
+                    cloudToDeviceMethod,
+                    timeout,
+                    null,
+                    null,
+                    cancellationToken);
+            }
+            catch (Exception ex)
+            {
+                Logging.Error(this, $"{nameof(InvokeDeviceMethodAsync)} threw an exception: {ex}", nameof(InvokeDeviceMethodAsync));
+                throw;
+            }
+            finally
+            {
+                Logging.Exit(this, $"Invoking device method for: {uri}", nameof(InvokeDeviceMethodAsync));
+            }
+        }
+
+        private static TimeSpan GetInvokeDeviceMethodOperationTimeout(CloudToDeviceMethod cloudToDeviceMethod)
+        {
+            // For InvokeDeviceMethod, we need to take into account the timeouts specified
+            // for the Device to connect and send a response. We also need to take into account
+            // the transmission time for the request send/receive
+            var timeout = TimeSpan.FromSeconds(15); // For wire time
+            timeout += TimeSpan.FromSeconds(cloudToDeviceMethod.ConnectionTimeoutInSeconds ?? 0);
+            timeout += TimeSpan.FromSeconds(cloudToDeviceMethod.ResponseTimeoutInSeconds ?? 0);
+            return timeout <= s_defaultOperationTimeout ? s_defaultOperationTimeout : timeout;
+        }
+
+        private static Uri GetStatisticsUri()
+        {
+            return new Uri(_statisticsUriFormat, UriKind.Relative);
+        }
+
+        private static Uri GetPurgeMessageQueueAsyncUri(string deviceId)
+        {
+            return new Uri(_purgeMessageQueueFormat.FormatInvariant(deviceId), UriKind.Relative);
+        }
+
+        private static Uri GetDeviceMethodUri(string deviceId)
+        {
+            deviceId = WebUtility.UrlEncode(deviceId);
+            return new Uri(_deviceMethodUriFormat.FormatInvariant(deviceId), UriKind.Relative);
+        }
+
+        private static Uri GetModuleMethodUri(string deviceId, string moduleId)
+        {
+            deviceId = WebUtility.UrlEncode(deviceId);
+            moduleId = WebUtility.UrlEncode(moduleId);
+            return new Uri(_moduleMethodUriFormat.FormatInvariant(deviceId, moduleId), UriKind.Relative);
+        }
     }
 }
